Sistemato metodo countByImmatricolateDopoAnno di Proprietario

+++ src/it/gestioneautomobili/dao/proprietario/ProprietarioDAO.java
@@ -4,5 +4,5 @@ import it.gestioneautomobili.dao.IBaseDAO;
 import it.gestioneautomobili.model.Proprietario;
 
 public interface ProprietarioDAO extends IBaseDAO<Proprietario> {
-	public int countByImmatricolateDopoAnno(int annoImmatricolazioneInput) throws Exception;
+	public Long countByImmatricolateDopoAnno(int annoImmatricolazioneInput) throws Exception;
 }
