+++ src/it/gestioneautomobili/service/proprietario/ProprietarioService.java
@@ -16,7 +16,7 @@ public interface ProprietarioService {
 
 	public void rimuovi(Proprietario proprietarioInstance) throws Exception;
 
-	public int quantiProprietariConAnnoImmatricolazioneMaggioreDi(int annoImmatricolazioneInput) throws Exception;
+	public Long quantiProprietariConAnnoImmatricolazioneMaggioreDi(int annoImmatricolazioneInput) throws Exception;
 
 	// per injection
 	public void setProprietarioDAO(ProprietarioDAO proprietarioDAO);
